test: run import stack checks under bun
